% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addColor.r
\name{addColor}
\alias{addColor}
\title{Add Color Information.}
\usage{
addColor(
  data,
  kit = NA,
  have = NA,
  need = NA,
  overwrite = FALSE,
  ignore.case = FALSE,
  debug = FALSE
)
}
\arguments{
\item{data}{data frame or vector.}

\item{kit}{string representing the forensic STR kit used.  Default is NA,
in which case 'have' must contain a valid column name.}

\item{have}{character string specifying which colour scheme is present
in the data.  Acceptable values are 'Color', 'Dye' or 'R.Color' (case
insensitive).  If data is a vector 'have' must be supplied.}

\item{need}{character string or vector specifying which colour schemes
should be added.  Acceptable values are 'Color', 'Dye' and 'R.Color'.
Default NA means that all missing colour columns will be added.
If data is a vector 'need' must be a single value.}

\item{overwrite}{logical.  If TRUE and the column to be created already
exists in the data frame it will be removed before adding new values.}

\item{ignore.case}{logical.  If TRUE marker names will be matched ignoring
case when deriving colours from a kit.}

\item{debug}{logical indicating whether to print debug information.}
}
\value{
For data frames, returns a data frame with additional columns
  corresponding to the requested colour schemes.  For vectors, returns a
  vector converted to the requested scheme.
}
\description{
Add color information 'Color', 'Dye' or 'R Color' to either a vector
or a data frame.  The function supports up to eight fluorescent dye
channels corresponding to modern STR kits and will gracefully handle
older kits with fewer dyes.
}
\details{
Primers in forensic STR typing kits are labelled with fluorescent dyes.
The dyes are represented with single letters (Dye) in exported result files,
with strings (Color) in 'panels' files and with R colour names for plotting
(R.Color).  This function can add missing colour columns to a data frame
based on another colour column, convert between schemes, or derive colours
from a kit definition via `getKit()`.  Conversions are case insensitive and
missing or unrecognised values will be mapped to NA.
}
