% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addData_gui}
\alias{addData_gui}
\title{Add data}
\usage{
addData_gui(env = parent.frame(), savegui = NULL, debug = FALSE,
  parent = NULL)
}
\arguments{
\item{env}{environment in wich to search for data frames.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
\code{addData_gui} is a GUI wrapper for \code{\link{addData}}.
}
\details{
Add new information to a dataset from a second dataset by
a key column.
}
\seealso{
\code{\link{addData}}
}

