\name{addSize_gui}
\alias{addSize_gui}
\title{Add size}
\usage{
addSize_gui(env = parent.frame(), savegui = NULL, debug = FALSE)
}
\arguments{
  \item{env}{environment in wich to search for data frames
  and save result.}

  \item{savegui}{logical indicating if GUI settings should
  be saved in the environment.}

  \item{debug}{logical indicating printing debug
  information.}
}
\value{
TRUE
}
\description{
\code{addSize_gui} is a GUI wrapper for the \code{\link{addSize}} function.
}
\details{
Simplifies the use of the \code{\link{addSize}} function by providing a
graphical user interface to it.
}
\seealso{
\code{\link{addSize}}
}

