% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculateAllele.R
\name{calculateAllele}
\alias{calculateAllele}
\title{Calculate Allele}
\usage{
calculateAllele(data, threshold = NULL, debug = FALSE)
}
\arguments{
\item{data}{data.frame including colums 'Marker', 'Allele', 'Height'.}

\item{threshold}{numeric if not NULL only peak heights above 'threshold'
will be considered.}

\item{debug}{logical indicating printing debug information.}
}
\value{
data.frame
}
\description{
Counts the number of each allele per marker over the entire dataset.
}
\details{
Creates a sorted table of the most common alleles in the dataset.
The list can be used to calculate allele frequencies or to identify artefacts.
NB! Remove NA's and OL's prior to analysis.
}
\seealso{
\code{\link{data.table}}
}

