% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateAllT_gui.r
\name{calculateAllT_gui}
\alias{calculateAllT_gui}
\title{Calculate Stochastic Thresholds}
\usage{
calculateAllT_gui(env = parent.frame(), savegui = NULL,
  debug = FALSE, parent = NULL)
}
\arguments{
\item{env}{environment in which to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI wrapper to the \code{\link{calculateAllT}} function.
}
\details{
Convenience GUI for the use of \code{\link{calculateAllT}} to calculate
point estimates for the stochastic threshold using multiple models.
}
\seealso{
\code{\link{calculateAllT}}
}
