% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateAT6_gui.r
\name{calculateAT6_gui}
\alias{calculateAT6_gui}
\title{Calculate Analytical Threshold}
\usage{
calculateAT6_gui(
  env = parent.frame(),
  savegui = NULL,
  debug = FALSE,
  parent = NULL
)
}
\arguments{
\item{env}{environment in which to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI wrapper for the \code{\link{calculateAT6}} function.
}
\details{
Scores dropouts for a dataset.
}
\seealso{
\code{\link{calculateAT6}}, \code{\link{calculateAT}},
 \code{\link{calculateAT_gui}}, \code{\link{checkSubset}}
}
