% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateAT_gui.r
\name{calculateAT_gui}
\alias{calculateAT_gui}
\title{Calculate Analytical Threshold}
\usage{
calculateAT_gui(
  env = parent.frame(),
  savegui = NULL,
  debug = FALSE,
  parent = NULL
)
}
\arguments{
\item{env}{environment in which to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI wrapper for the \code{\link{maskAT}} and \code{\link{calculateAT}} function.
}
\details{
Simplifies the use of the \code{\link{calculateAT}} and
 \code{\link{calculateAT}} function by providing a graphical user interface.
 In addition there are integrated control functions.
}
\seealso{
\code{\link{calculateAT}}, \code{\link{maskAT}},
 \code{\link{checkSubset}}
}
