% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateBalance_gui.r
\name{calculateBalance_gui}
\alias{calculateBalance_gui}
\title{Calculate Balance}
\usage{
calculateBalance_gui(env = parent.frame(), savegui = NULL, debug = FALSE,
  parent = NULL)
}
\arguments{
\item{env}{environment in wich to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI wrapper for the \code{\link{calculateBalance}} function.
}
\details{
Simplifies the use of the \code{\link{calculateBalance}} function
by providing a graphical user interface.
}
\seealso{
\code{link{calculateBalance}}, \code{link{checkSubset}}
}

