% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateHb_gui.r
\name{calculateHb_gui}
\alias{calculateHb_gui}
\title{Calculate Heterozygote Balance}
\usage{
calculateHb_gui(env = parent.frame(), savegui = NULL, debug = FALSE,
  parent = NULL)
}
\arguments{
\item{env}{environment in wich to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI wrapper for the \code{\link{calculateHb}} function.
}
\details{
Simplifies the use of the \code{\link{calculateHb}} function
by providing a graphical user interface.
}
\seealso{
\code{link{calculateHb}}, \code{link{checkSubset}}
}

