% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateHeight_gui.r
\name{calculateHeight_gui}
\alias{calculateHeight_gui}
\title{Calculate Peak Height}
\usage{
calculateHeight_gui(env = parent.frame(), savegui = NULL,
  debug = FALSE, parent = NULL)
}
\arguments{
\item{env}{environment in which to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI wrapper for the \code{\link{calculateHeight}} function.
}
\details{
Simplifies the use of the \code{\link{calculateHeight}} function by providing a graphical
user interface to it.
}
\references{
Torben Tvedebrink, Poul Svante Eriksen, Helle Smidt Mogensen, Niels Morling,
 Evaluating the weight of evidence by using quantitative short tandem repeat data in DNA mixtures
 Journal of the Royal Statistical Society: Series C (Applied Statistics),
 Volume 59, Issue 5, 2010,
 Pages 855-874, 10.1111/j.1467-9876.2010.00722.x.
\url{http://dx.doi.org/10.1111/j.1467-9876.2010.00722.x}
}
\seealso{
\code{\link{calculateHeight}}
}
