% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMixture_gui.r
\name{calculateMixture_gui}
\alias{calculateMixture_gui}
\title{Calculate Mixture}
\usage{
calculateMixture_gui(
  env = parent.frame(),
  savegui = NULL,
  debug = FALSE,
  parent = NULL
)
}
\arguments{
\item{env}{environment in which to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI wrapper for the \code{\link{calculateMixture}} function.
}
\details{
Simplifies the use of the \code{\link{calculateMixture}} function by
providing a graphical user interface.
}
\seealso{
\code{\link{calculateMixture}}, \code{\link{checkSubset}}
}
