% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculatePullup_gui.r
\name{calculatePullup_gui}
\alias{calculatePullup_gui}
\title{Calculate Spectral Pull-up}
\usage{
calculatePullup_gui(env = parent.frame(), savegui = NULL, debug = FALSE,
  parent = NULL)
}
\arguments{
\item{env}{environment in wich to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI wrapper for the \code{\link{calculatePullup}} function.
}
\details{
Simplifies the use of the \code{\link{calculatePullup}} function by
providing a graphical user interface.
}
\seealso{
\code{\link{calculatePullup}}, \code{\link{checkSubset}}
}

