% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRatio_gui.r
\name{calculateRatio_gui}
\alias{calculateRatio_gui}
\title{Calculate Ratio}
\usage{
calculateRatio_gui(env = parent.frame(), savegui = NULL, debug = FALSE,
  parent = NULL)
}
\arguments{
\item{env}{environment in which to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI wrapper for the \code{\link{calculateRatio}} function.
}
\details{
Simplifies the use of the \code{\link{calculateRatio}} function
by providing a graphical user interface.
}
\seealso{
\code{link{calculateRatio}}, \code{link{checkSubset}}
}
