% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSpike_gui.r
\name{calculateSpike_gui}
\alias{calculateSpike_gui}
\title{Detect Spike}
\usage{
calculateSpike_gui(env = parent.frame(), savegui = NULL, debug = FALSE,
  parent = NULL)
}
\arguments{
\item{env}{environment in wich to search for data frames.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI wrapper for the \code{\link{calculateSpike}} function.
}
\details{
Simplifies the use of the \code{\link{calculateSpike}} function
by providing a graphical user interface.
}
\seealso{
\code{\link{calculateSpike}}
}

