% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDataset.r
\name{checkDataset}
\alias{checkDataset}
\title{Check Dataset}
\usage{
checkDataset(name, reqcol = NULL, slim = FALSE, slimcol = NULL,
  string = NULL, stringcol = NULL, env = parent.frame(),
  parent = NULL, debug = FALSE)
}
\arguments{
\item{name}{character name of data.frame.}

\item{reqcol}{character vector with required column names.}

\item{slim}{logical TRUE to check if 'slim' data.}

\item{slimcol}{character vector with column names to check if 'slim' data.}

\item{string}{character vector with invalid strings in 'stringcol', return FALSE if found.}

\item{stringcol}{character vector with column names to check for 'string'.}

\item{env}{environment where to look for the data frame.}

\item{parent}{parent gWidget.}

\item{debug}{logical indicating printing debug information.}
}
\description{
Internal function to check a data.frame before analysis.
}
\details{
Check that the object exist, there are rows, the required columns exist,
if data.frame is 'fat', and if invalid strings exist. Show error message if not.
}
