% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colConvert.r
\name{colConvert}
\alias{colConvert}
\title{Convert Columns}
\usage{
colConvert(data, columns = "Height|Size|Data.Point", ignore.case = TRUE,
  fixed = FALSE, debug = FALSE)
}
\arguments{
\item{data}{data.frame.}

\item{columns}{character string containing a regular expression
(or character string for fixed = TRUE) to be matched in the given
character vector (separate multiple column names by | in reg.exp).}

\item{ignore.case}{logical TRUE to ignore case in matching.}

\item{fixed}{logical TRUE if columns is a string to be matched as is.}

\item{debug}{logical indicating printing debug information.}
}
\value{
data.frame.
}
\description{
Internal helper function.
}
\details{
Takes a data frame as input and return it after converting known numeric
columns to numeric.
}

