% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns_gui.r
\name{columns_gui}
\alias{columns_gui}
\title{Column Actions}
\usage{
columns_gui(env = parent.frame(), savegui = NULL, debug = FALSE,
  parent = NULL)
}
\arguments{
\item{env}{environment in which to search for data frames.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI wrapper for the \code{\link{columns}} function.
}
\details{
Simplifies the use of the \code{\link{columns}} function by providing a
 graphical user interface to it.
}
