% Generated by roxygen2 (4.0.0): do not edit by hand
\name{concatenate_gui}
\alias{concatenate_gui}
\title{Concatenate two datasets}
\usage{
concatenate_gui(env = parent.frame(), debug = FALSE)
}
\arguments{
\item{env}{environment in wich to search for data frames.}

\item{debug}{logical indicating printing debug information.}
}
\description{
\code{concatenate_gui} is a GUI for combining two datasets.
}
\details{
Simplifies the use of the \code{checkSubset} function by providing a graphical
user interface to it.
}

