\name{countDropin}
\alias{countDropin}
\title{Calculate drop-in peaks}
\usage{
  countDropin(data, col = "Height", unique = TRUE)
}
\arguments{
  \item{data}{a data frame containing at least
  'Sample.Name'.}
}
\value{
  data.frame data frame with
  'Sample.Name','Marker','Allele','Peaks')
}
\description{
  \code{calculateDropin} calculates number of drop-in peaks
}
\details{
  Not sure if this function is working (might be a copy of
  countPeaks as a draft...)
}
\examples{
print("Example will come later")
}
\keyword{internal}

