\name{editData_gui}
\alias{editData_gui}
\title{Edit or view data frames GUI}
\usage{
  editData_gui(env = parent.frame(), debug = FALSE)
}
\arguments{
  \item{env}{environment in wich to search for data
  frames.}

  \item{debug}{logical indicating printing debug
  information.}
}
\description{
  \code{editData_gui} is a GUI simplifying editing and
  viewing of data frames.
}
\details{
  Select a data frame from the dropdown and view/edit.
  Optionally save as a new dataframe.
}

