% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_gui.r
\name{export_gui}
\alias{export_gui}
\title{Export}
\usage{
export_gui(obj = listObjects(env = env, obj.class = c("data.frame",
  "ggplot")), env = parent.frame(), savegui = NULL, debug = FALSE,
  parent = NULL)
}
\arguments{
\item{obj}{character vector with object names.}

\item{env}{environment where the objects exist.
Default is the current environment.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI wrapper for the \code{\link{export}} function.
}
\details{
Simplifies the use of the \code{\link{export}} function by providing a 
graphical user interface to it. Currently all available objects provided
are selected by default.
}
\seealso{
\code{\link{export}}
}
