% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSetting.r
\name{getSetting}
\alias{getSetting}
\title{Get Settings.}
\usage{
getSetting(key)
}
\arguments{
\item{key}{character key for value to return.}
}
\value{
character the retrieved value or NA if not found.
}
\description{
Accepts a key string and returns the corresponding value.
}
\details{
Accepts a key string and returns the corresponding value from the settings.txt
file located within the package folders exdata sub folder.
}
