\name{guessProfile_gui}
\alias{guessProfile_gui}
\title{Guess profile}
\usage{
  guessProfile_gui(env = parent.frame(), debug = FALSE)
}
\arguments{
  \item{env}{environment in wich to search for data
  frames.}

  \item{debug}{logical indicating printing debug
  information.}
}
\description{
  \code{guessProfile_gui} is a GUI wrapper for the
  \code{guessProfile} function.
}
\details{
  Simplifies the use of the \code{guessProfile} function by
  providing a graphical user interface to it.
}

