% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heightToPeak.r
\name{heightToPeak}
\alias{heightToPeak}
\title{Height To Peak.}
\usage{
heightToPeak(data, width = 1, keep.na = TRUE, debug = FALSE)
}
\arguments{
\item{data}{data frame containing at least columns 'Height' and 'Size'.}

\item{width}{numeric specifying the width of the peak in bp.}

\item{keep.na}{logical. TRUE to keep empty markers.}

\item{debug}{logical. TRUE prints debug information.}
}
\value{
data.frame with new values.
}
\description{
Helper function to convert a peak into a plotable polygon.
}
\details{
Converts a single height and size value to a plotable 0-height-0 triangle/peak value.
Makes 3 data points from each peak size for plotting a polygon representing a peak.
Factors in other columns might get converted to factor level.
}
