% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeKit_gui.r
\name{makeKit_gui}
\alias{makeKit_gui}
\title{Make Kit}
\usage{
makeKit_gui(env = parent.frame(), savegui = NULL, debug = FALSE, parent = NULL)
}
\arguments{
\item{env}{environment in which to search for data frames.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.
[Not currently used]}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
Add new kits or edit the kit file.
}
\details{
A graphical user interface for reading information from 'bins' and
'panels' file for the creation of additional kits. It is also possible to
edit the short and full name of existing kits or removing kits.
The gender marker of each kits is auto detected but can be changed manually.
#' NB! Short name must be unique.
}
\seealso{
\code{\link{readBinsFile}}, \code{\link{readPanelsFile}}, \code{\link{combineBinsAndPanels}}
}
