% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageKits_gui.r
\name{manageKits_gui}
\alias{manageKits_gui}
\title{Manage Kits}
\usage{
manageKits_gui(
  env = parent.frame(),
  savegui = NULL,
  debug = FALSE,
  parent = NULL
)
}
\arguments{
\item{env}{Environment in which to search for data frames.}

\item{savegui}{Logical indicating if GUI settings should be saved in the
environment (Currently not in use).}

\item{debug}{Logical indicating whether to print debug information.}

\item{parent}{Widget to get focus when finished.}
}
\value{
TRUE if the operation completes successfully.
}
\description{
Manage kits, import new kits, or edit the kit file through a graphical user
 interface.
}
\details{
This function provides a graphical user interface (GUI) for managing kits,
including the ability to import new kits, edit the short and full names of
existing kits, or remove kits. The gender marker of each kit is auto-detected
but can be manually adjusted. Note that the short name for each kit must be
unique.
}
