\name{plotBalance_gui}
\alias{plotBalance_gui}
\title{Plot Balance GUI}
\usage{
  plotBalance_gui(env = parent.frame(), debug = FALSE)
}
\arguments{
  \item{env}{environment in wich to search for data frames
  and save result.}

  \item{debug}{logical indicating printing debug
  information.}
}
\description{
  \code{plotBalance_gui} is a GUI simplifying the creation
  of plots from balance data.
}
\details{
  Plot balance data.
}

