% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plotCapillary_gui}
\alias{plotCapillary_gui}
\title{Plot Capillary Balance GUI}
\usage{
plotCapillary_gui(env = parent.frame(), savegui = NULL, debug = FALSE)
}
\arguments{
\item{env}{environment in wich to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}
}
\description{
\code{plotCapillary_gui} is a GUI simplifying the creation of plots from
capillary balance data.
}
\details{
Plot capillary balance data.
}

