% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotContamination_gui.r
\name{plotContamination_gui}
\alias{plotContamination_gui}
\title{Plot Contamination}
\usage{
plotContamination_gui(env = parent.frame(), savegui = NULL, debug = FALSE,
  parent = NULL)
}
\arguments{
\item{env}{environment in wich to search for data frames.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI simplifying the creation of plots from negative control data.
}
\details{
Select data to plot in the drop-down menu. 
Automatic plot titles can be replaced by custom titles.
A name for the result is automatiaclly suggested.
The resulting plot can be saved as either a plot object or as an image.
}
\references{
Duncan Taylor et.al.,
 Validating multiplexes for use in conjunction with modern interpretation strategies,
 Forensic Science International: Genetics, Volume 20, January 2016,
 Pages 6-19, ISSN 1872-4973, 10.1016/j.fsigen.2015.09.011.
\url{http://www.sciencedirect.com/science/article/pii/S1872497315300739}
}

