\name{plotDistribution_gui}
\alias{plotDistribution_gui}
\title{Plot Distribution GUI}
\usage{
plotDistribution_gui(env = parent.frame(), savegui = NULL, debug = FALSE)
}
\arguments{
  \item{env}{environment in wich to search for data frames
  and save result.}

  \item{savegui}{logical indicating if GUI settings should
  be saved in the environment.}

  \item{debug}{logical indicating printing debug
  information.}
}
\description{
\code{plotDistribution_gui} is a GUI simplifying the
creation of distribution plots.
}
\details{
Plot result type data.
}

