% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDropout_gui.r
\name{plotDropout_gui}
\alias{plotDropout_gui}
\title{Plot Drop-out Events}
\usage{
plotDropout_gui(
  env = parent.frame(),
  savegui = NULL,
  debug = FALSE,
  parent = NULL
)
}
\arguments{
\item{env}{environment in which to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI simplifying the creation of plots from dropout data.
}
\details{
Plot dropout data as heatmap arranged by, average peak height,
amount, concentration, or sample name. It is also possible to plot the
empirical cumulative distribution (ecdp) of the peak heights of surviving heterozygote
alleles (with dropout of the partner allele), or a dotplot of all dropout events.
The peak height of homozygote alleles can be included in the ecdp.
Automatic plot titles can be replaced by custom titles.
A name for the result is automatically suggested.
The resulting plot can be saved as either a plot object or as an image.
}
\references{
Antoinette A. Westen, Laurens J.W. Grol, Joyce Harteveld, Anuska S.Matai,
Peter de Knijff, Titia Sijen, Assessment of the stochastic threshold, back- and
forward stutter filters and low template techniques for NGM,
Forensic Science International: Genetetics, Volume 6, Issue 6, December 2012,
Pages 708-715, ISSN 1872-4973, 10.1016/j.fsigen.2012.05.001.
 \doi{10.1016/j.fsigen.2012.05.001}
}
\seealso{
\url{https://ggplot2.tidyverse.org/} for details on plot settings.
}
