% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEPG2_gui.r
\name{plotEPG2_gui}
\alias{plotEPG2_gui}
\title{Plot EPG}
\usage{
plotEPG2_gui(
  env = parent.frame(),
  savegui = NULL,
  debug = FALSE,
  parent = NULL
)
}
\arguments{
\item{env}{environment in which to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI wrapper for the \code{\link{plotEPG2}} function.
}
\details{
Simplifies the use of the \code{\link{plotEPG2}} function by providing a graphical
user interface.
}
\seealso{
\code{\link{plotEPG2}}
}
