% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGroups_gui.r
\name{plotGroups_gui}
\alias{plotGroups_gui}
\title{Plot Empirical Cumulative Distributions}
\usage{
plotGroups_gui(
  env = parent.frame(),
  savegui = NULL,
  debug = FALSE,
  parent = NULL
)
}
\arguments{
\item{env}{environment in which to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI simplifying the creation of empirical cumulative distribution plots.
}
\details{
Plot the distribution of data as cumulative distribution function
for multiple groups. First select a dataset, then select columns to flat,
group, and plot by. For example, if a genotype dataset is selected and data
is flattened by Sample.Name the 'group by' and 'plot by' values must be
identical for all rows for a given sample.
Automatic plot titles can be replaced by custom titles.
Group names can be changed. A name for the result is automatically suggested.
The resulting plot can be saved as either a plot object or as an image.
}
\seealso{
\code{\link{stat_ecdf}}
}
