\name{plotPeaks_gui}
\alias{plotPeaks_gui}
\title{Plot Peaks GUI}
\usage{
plotPeaks_gui(env = parent.frame(), savegui = NULL, debug = FALSE)
}
\arguments{
  \item{env}{environment in wich to search for data frames
  and save result.}

  \item{savegui}{logical indicating if GUI settings should
  be saved in the environment.}

  \item{debug}{logical indicating printing debug
  information.}
}
\description{
\code{plotPeaks_gui} is a GUI simplifying the creation of
plots from result type data.
}
\details{
Plot result type data.
}

