% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRatio_gui.r
\name{plotRatio_gui}
\alias{plotRatio_gui}
\title{Plot Ratio}
\usage{
plotRatio_gui(env = parent.frame(), savegui = NULL, debug = FALSE,
  parent = NULL)
}
\arguments{
\item{env}{environment in which to search for data frames.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI simplifying the creation of plots from marker ratio data.
}
\details{
Select data to plot in the drop-down menu. 
Automatic plot titles can be replaced by custom titles.
A name for the result is automatically suggested.
The resulting plot can be saved as either a plot object or as an image.
}
\seealso{
\url{http://docs.ggplot2.org/current/} for details on plot settings.
}
