% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gene_mapper_kit.r
\name{read_gene_mapper_kit}
\alias{read_gene_mapper_kit}
\title{Read GeneMapper Kit Definition}
\usage{
read_gene_mapper_kit(bin_files = NULL, panel_files = NULL, debug = FALSE)
}
\arguments{
\item{parent}{widget to get focus when finished.}
}
\value{
data.frame
}
\description{
Import kit definition from GeneMapper bins and panel files.
}
\details{
Takes the GeneMapper bins and panels file and creates a
kit definition data frame.
}
