% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gene_marker_kit.r
\name{read_gene_marker_kit}
\alias{read_gene_marker_kit}
\title{Read GeneMarker Kit Definition}
\usage{
read_gene_marker_kit(xml_file_path, panel_name)
}
\arguments{
\item{xml_file_path}{the path to the XML file.}

\item{panel_name}{the name of the panel to be imported.}
}
\value{
data.frame
}
\description{
Import kit definition from GeneMarker XML-files.
}
\details{
Takes the GeneMarker kit XML-file and creates a
kit definition data frame.
}
