% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gene_marker_kit_gui.r
\name{read_gene_marker_kit_gui}
\alias{read_gene_marker_kit_gui}
\title{Read GeneMarker Kit Definition}
\usage{
read_gene_marker_kit_gui(
  env = parent.frame(),
  savegui = NULL,
  debug = FALSE,
  parent = NULL,
  callback = NULL
)
}
\arguments{
\item{env}{environment in which to search for data frames.}

\item{savegui}{logical indicating if GUI settings should be saved in
the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
data.frame
}
\description{
Read GeneMarker kit definition file.
}
\details{
Select the kit definition XML-file using the file picker.
The unique Panel Names will appear in the drop-down menu.
Select the panel to import.
Kit information for the selected panel will be extracted.
}
