% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBinsFile.r
\name{readBinsFile}
\alias{readBinsFile}
\title{Read Bins file}
\usage{
readBinsFile(bin.files, debug = FALSE)
}
\arguments{
\item{bin.files}{string, complete path to Bins file.}

\item{debug}{logical indicating printing debug information.}
}
\value{
data.frame containing the columns 'Panel', 'Marker', 'Allele', 'Size',
'Size.Min', 'Size.Max', 'Virtual'.
}
\description{
Reads GeneMapper 'Bins' files.
}
\details{
Reads useful information from 'Bins' files and save it as a data.frame.
}
\keyword{internal}

