% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGeneMapperKit.r
\name{readGeneMapperKit}
\alias{readGeneMapperKit}
\title{Read GeneMapper Kit Definition}
\usage{
readGeneMapperKit(bin.files = NULL, panel.files = NULL, debug = FALSE)
}
\arguments{
\item{parent}{widget to get focus when finished.}
}
\value{
data.frame
}
\description{
Import kit definition from GeneMapper bins and panel files.
}
\details{
Takes the GeneMapper bins and panels file and creates a
kit definition data frame.
}
