% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGeneMapperKit_gui.r
\name{readGeneMapperKit_gui}
\alias{readGeneMapperKit_gui}
\title{Read GeneMapper Kit Definition (GUI)}
\usage{
readGeneMapperKit_gui(
  env = globalenv(),
  savegui = TRUE,
  debug = FALSE,
  parent = NULL,
  callback = NULL
)
}
\arguments{
\item{env}{environment in which to search for data frames.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
data.frame
}
\description{
Import GeneMapper kit definition files through a graphical user interface.
}
\details{
Select the kit bins and panels file using the file picker.
}
