% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGeneMarkerKit.r
\name{readGeneMarkerKit}
\alias{readGeneMarkerKit}
\title{Read GeneMarker Kit Definition}
\usage{
readGeneMarkerKit(xml.file.path, panel.name)
}
\arguments{
\item{xml.file.path}{the path to the XML file.}

\item{panel.name}{the name of the panel to be imported.}
}
\value{
data.frame
}
\description{
Import kit definition from GeneMarker XML-files.
}
\details{
Takes the GeneMarker kit XML-file and creates a
kit definition data frame.
}
