% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strvalidator-package.r
\docType{data}
\name{ref1}
\alias{ref1}
\title{ESX17 Positive Control Profile}
\format{A data frame with 17 rows and 4 variables}
\usage{
data(ref1)
}
\description{
A dataset in 'GeneMaper' format containing the DNA profile of
the ESX17 positive control sample with homozygotes as one entry.
}
\keyword{datasets}

