% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strvalidator-package.r
\docType{data}
\name{ref11}
\alias{ref11}
\title{ESX17 Positive Control Profile}
\format{A data frame with 17 rows and 4 variables}
\usage{
data(ref11)
}
\description{
A dataset in 'GeneMaper' format containing the DNA profile of
the ESX17 positive control sample with homozygotes as two entries.
}
\keyword{datasets}

