% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strvalidator-package.r
\docType{data}
\name{ref2}
\alias{ref2}
\title{SGMPlus example data}
\format{A data frame with 16 rows and 3 variables}
\usage{
data(ref2)
}
\description{
A slimmed reference dataset containing an arbitrary SGMPlus DNA profile.
}
\keyword{datasets}
