% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strvalidator-package.r
\name{ref3}
\alias{ref3}
\title{ESX17 Example Data for Dropout Analysis}
\format{
ASCII text file
}
\description{
Reference profiles for source samples.
Text file in GeneMapper format.
}
\keyword{datasets}
