% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strvalidator-package.r
\docType{data}
\name{ref4}
\alias{ref4}
\title{ESX17 example data for dropout analysis.}
\format{
A data frame with 98 rows and 3 variables
}
\usage{
data(ref4)
}
\description{
A slimmed dataset containing reference profiles for source samples in set4.
Reference 'A2' has double entries for homozygotes.
Reference 'F2' has single entries for homozygotes.
Reference 'bc' has double entries for homozygotes, and lower case sample name.
}
\keyword{datasets}
