% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/strvalidator-package.r
\docType{data}
\name{ref51}
\alias{ref51}
\title{ESX17 example data for mixture analysis.}
\format{A data frame with 34 rows and 3 variables}
\usage{
data(ref51)
}
\description{
A slimmed dataset containing the reference profile for the major
component  in set5.
}
\keyword{datasets}

