% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ref52}
\alias{ref52}
\title{ESX17 example data for mixture analysis.}
\format{A data frame with 34 rows and 3 variables}
\usage{
data(ref52)
}
\description{
A slimmed dataset containing the reference profile for the minor
component  in set5.
}
\keyword{datasets}

