% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strvalidator-package.r
\name{ref62}
\alias{ref62}
\title{Fusion Example Data for Dropout Analysis}
\format{
A data frame with 89 rows and 3 variables
}
\usage{
data(ref62)
}
\description{
A slimmed dataset containing the reference profile for the
samples in set6.
NB! Marker order is the same as set6.
NB! Reference R has a Y marker with NA.
}
\keyword{datasets}
