% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strvalidator-package.r
\docType{data}
\name{ref7}
\alias{ref7}
\title{ESSplex SE QS example data for inhibition analysis.}
\format{
A data frame with 35 rows and 4 variables
}
\usage{
data(ref7)
}
\description{
A slimmed dataset containing the reference profile for the
samples in set7.
}
\keyword{datasets}
