% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strvalidator-package.r
\name{set1}
\alias{set1}
\title{Typing Data in 'GeneMapper' Format}
\format{
A data frame with 170 rows and 13 variables
}
\usage{
data(set1)
}
\description{
A dataset containing ESX17 genotyping results for 8 replicates
of the positive control sample, a negative control, and a ladder.
}
\keyword{datasets}
