% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strvalidator-package.r
\name{set3}
\alias{set3}
\title{ESX17 Example Data for Dropout Analysis}
\format{
ASCII text file
}
\description{
Data from a dilution experiment for dropout analysis.
Text file with exported GeneMapper genotypes table.
}
\keyword{datasets}
