% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strvalidator-package.r
\name{set4}
\alias{set4}
\title{ESX17 Example Data for Dropout Analysis}
\format{
A data frame with 1609 rows and 5 variables
}
\usage{
data(set4)
}
\description{
A slimmed dataset containing data from
a dilution experiment for dropout analysis (from set3).
One sample replicate has a lower case sample name (bc9).
}
\keyword{datasets}
