% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strvalidator-package.r
\name{set5}
\alias{set5}
\title{ESX17 Example Data for Mixture Analysis}
\format{
A data frame with 1663 rows and 7 variables
}
\usage{
data(set5)
}
\description{
A slimmed dataset containing data from
a mixture experiment for Mx analysis.
}
\keyword{datasets}
