% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strvalidator-package.r
\name{set6}
\alias{set6}
\title{Fusion Example Data for Dropout Analysis}
\format{
A data frame with 1848 rows and 7 variables
}
\usage{
data(set6)
}
\description{
A slimmed dataset containing data from
a sensitivity experiment for dropout analysis.
}
\keyword{datasets}
