% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strvalidator-package.r
\name{set7}
\alias{set7}
\title{ESSplex SE QS Example Data for Inhibition Analysis}
\format{
A data frame with 883 rows and 7 variables
}
\usage{
data(set7)
}
\description{
A slimmed dataset containing data from
an inhibition experiment.
}
\keyword{datasets}
