% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortMarker.r
\name{sortMarker}
\alias{sortMarker}
\title{Sort Markers}
\usage{
sortMarker(data, kit, add.missing.levels = FALSE, debug = FALSE)
}
\arguments{
\item{data}{data.frame containing a column 'Marker' and optionally 'Dye'.}

\item{kit}{string or integer indicating kit.}

\item{add.missing.levels}{logical, TRUE missing markers are added, 
FALSE missing markers are not added.}

\item{debug}{logical indicating printing debug information.}
}
\value{
data.frame with factor levels sorted according to 'kit'.
}
\description{
Sort markers and dye as they appear in the EPG.
}
\details{
Change the order of factor levels for 'Marker' and 'Dye' according to 'kit'.
Levels in data must be identical with kit information.
}

