\docType{package}
\name{strValidator-package}
\alias{strValidator}
\alias{strValidator-package}
\title{Simplifies internal validation of forensic STR typing kits}
\description{
  Internal validation of forensic STR kits made easy with
  strValidator.
}
\details{
  strValidator provide tools that make it easier to perform
  an internal validation of a forensic short tandem repeat
  (STR) kit for human identification. The tools are
  developed to provide all necessary data to conform with
  the ENFSI guidelines in "Recommended Minimum Criteria for
  the Validation of Various Aspects of the DNA Profiling
  Process - New Multiplex Kit" More information about each
  function can be found in its help documentation. The
  package is still in an early developmental stage and
  things will likely change in coming versions. Manually
  check the results thoroughly, the software has not been
  validated. Please report bugs to:
  https://github.com/OskarHansson/strvalidator/issues
}
\section{Warning}{
  This package is experimental and has not been thoroughly
  validated.
}
\author{
  Oskar Hansson \email{oskar.hansson@fhi.no}
}
\keyword{package}

