% Generated by roxygen2 (4.0.0): do not edit by hand
\name{strvalidator}
\alias{strvalidator}
\title{GUI for STR Validator}
\usage{
strvalidator(debug = FALSE)
}
\arguments{
\item{debug}{logical indicating printing debug information.}
}
\description{
\code{strvalidator} is validation toolbox GUI .
}
\details{
This graphical user interface make it very easy to perform an internal
validation of forensic STR kits. The GUI works as a toolbox for functions
required to analyse validation data.
}
\examples{
# To start the graphical user interface.
\dontrun{
strvalidator()
}
}

