\name{tableBalance_gui}
\alias{tableBalance_gui}
\title{Table balance}
\usage{
tableBalance_gui(env = parent.frame(), savegui = NULL, debug = FALSE)
}
\arguments{
  \item{env}{environment in wich to search for data
  frames.}

  \item{savegui}{logical indicating if GUI settings should
  be saved in the environment.}

  \item{debug}{logical indicating printing debug
  information.}
}
\description{
\code{tableBalance_gui} is a GUI wrapper for the
\code{tableBalance} function.
}
\details{
Simplifies the use of the \code{tableBalance} function by
providing a graphical user interface to it.
}

