\name{tableCapillary_gui}
\alias{tableCapillary_gui}
\title{Table Capillary}
\usage{
tableCapillary_gui(env = parent.frame(), savegui = NULL, debug = FALSE)
}
\arguments{
  \item{env}{environment in wich to search for data
  frames.}

  \item{savegui}{logical indicating if GUI settings should
  be saved in the environment.}

  \item{debug}{logical indicating printing debug
  information.}
}
\value{
TRUE
}
\description{
\code{tableCapillary_gui} is a GUI wrapper for the \code{\link{tableCapillary}} function.
}
\details{
Simplifies the use of the \code{\link{tableCapillary}} function by providing a graphical
user interface to it.
}
\seealso{
\code{\link{tableCapillary}}
}

