% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tablePrecision_gui}
\alias{tablePrecision_gui}
\title{Calculate Precision}
\usage{
tablePrecision_gui(env = parent.frame(), savegui = NULL, debug = FALSE,
  parent = NULL)
}
\arguments{
\item{env}{environment in wich to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
\code{tablePrecision_gui} is a GUI wrapper for the \code{\link{tablePrecision}}
 function.
}
\details{
Simplifies the use of the \code{\link{tablePrecision}} function by providing
a graphical user interface.
}
\seealso{
\code{\link{tablePrecision}}, \code{\link{checkSubset}}
}

