% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStrings.r
\name{update_strings_with_language_file}
\alias{update_strings_with_language_file}
\title{Update Strings with Language File}
\usage{
update_strings_with_language_file(default_strings, language_strings)
}
\arguments{
\item{default_strings}{list of default strings.}

\item{language_strings}{list of language-specific strings.}
}
\value{
list of updated strings.
}
\description{
Updates the default strings with the values from the language file.
}
