% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addData_gui.r
\name{addData_gui}
\alias{addData_gui}
\title{Add Data}
\usage{
addData_gui(env = parent.frame(), savegui = NULL, debug = FALSE,
  parent = NULL)
}
\arguments{
\item{env}{environment in wich to search for data frames.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI wrapper for \code{\link{addData}}.
}
\details{
Simplifies the use of the \code{\link{addData}} function by providing a graphical 
user interface to it.
}
\seealso{
\code{\link{addData}}
}

