% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDye_gui.r
\name{addDye_gui}
\alias{addDye_gui}
\title{Add Dye Information}
\usage{
addDye_gui(env = parent.frame(), savegui = NULL, debug = FALSE,
  parent = NULL)
}
\arguments{
\item{env}{environment in wich to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI wrapper to the \code{\link{addColor}} function.
}
\details{
Convenience GUI for the use of \code{\link{addColor}} and
\code{\link{addOrder}} to add 'Dye', 'Color', 'R.Color', and marker 'Order'
to a dataset.
'Dye' is the one letter abbreviations for the fluorophores commonly used
to label primers in forensic STR typing kits (e.g. R and Y),
'Color' is the corresponding color name (e.g. red and yellow),
'R.Color' is the plot color used in R (e.g. red and black).
'Order' is the marker order in the selected kit.
NB! Existing columns will be overwritten.
}
\seealso{
\code{\link{addColor}}
}
