% Generated by roxygen2 (4.0.0): do not edit by hand
\name{addDye_gui}
\alias{addDye_gui}
\title{Adds dye information}
\usage{
addDye_gui(env = parent.frame(), savegui = NULL, debug = FALSE)
}
\arguments{
\item{env}{environment in wich to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}
}
\value{
data.frame in slim format.
}
\description{
\code{addDye_gui} is a GUI wrapper for the \code{addDye} function.
}
\details{
Simplifies the use of the \code{addDye} function by providing a graphical
user interface to it.
}

