\name{addDye_gui}
\alias{addDye_gui}
\title{Adds dye information}
\usage{
addDye_gui(env = parent.frame(), debug = FALSE)
}
\arguments{
  \item{env}{environment in wich to search for data frames
  and save result.}

  \item{debug}{logical indicating printing debug
  information.}
}
\value{
data.frame in slim format.
}
\description{
\code{addDye_gui} is a GUI wrapper for the \code{addDye}
function.
}
\details{
Simplifies the use of the \code{addDye} function by
providing a graphical user interface to it.
}

