\name{addSize_gui}
\alias{addSize_gui}
\title{Adds size information}
\usage{
addSize_gui(env = parent.frame(), debug = FALSE)
}
\arguments{
  \item{env}{environment in wich to search for data frames
  and save result.}

  \item{debug}{logical indicating printing debug
  information.}
}
\value{
data.frame in slim format.
}
\description{
\code{addSize_gui} is a GUI wrapper for the \code{addSize}
function.
}
\details{
Simplifies the use of the \code{addSize} function by
providing a graphical user interface to it.
}

