% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSize_gui.r
\name{addSize_gui}
\alias{addSize_gui}
\title{Add Size Information}
\usage{
addSize_gui(env = parent.frame(), savegui = NULL, debug = FALSE, parent = NULL)
}
\arguments{
\item{env}{environment in which to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI wrapper for the \code{\link{addSize}} function.
}
\details{
Simplifies the use of the \code{\link{addSize}} function by providing a
graphical user interface to it.
}
\seealso{
\code{\link{addSize}}
}
