% Generated by roxygen2 (4.0.0): do not edit by hand
\name{calculateCapillary_gui}
\alias{calculateCapillary_gui}
\title{Capillary balance}
\usage{
calculateCapillary_gui(env = parent.frame(), savegui = NULL,
  debug = FALSE)
}
\arguments{
\item{env}{environment in wich to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}
}
\value{
data.frame in slim format.
}
\description{
\code{capillaryBalance_gui} is a GUI wrapper for the \code{capillaryBalance}
 function.
}
\details{
Simplifies the use of the \code{capillaryBalance} function by providing
a graphical user interface.
}

