% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculateConcordance_gui.r
\name{calculateConcordance_gui}
\alias{calculateConcordance_gui}
\title{Calculate Concordance}
\usage{
calculateConcordance_gui(env = parent.frame(), savegui = NULL,
  debug = FALSE, parent = NULL)
}
\arguments{
\item{env}{environment in wich to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI wrapper for the \code{\link{calculateConcordance}} function.
}
\details{
Simplifies the use of the \code{\link{calculateConcordance}} function by
providing a graphical user interface.
}
\seealso{
\code{\link{calculateConcordance}}
}

