\name{calculateDropout_gui}
\alias{calculateDropout_gui}
\title{Calculate Dropout GUI}
\usage{
  calculateDropout_gui(env = parent.frame(), debug = FALSE)
}
\arguments{
  \item{env}{environment in wich to search for data frames
  and save result.}

  \item{debug}{logical indicating printing debug
  information.}
}
\description{
  \code{calculateDropout_gui} is a GUI wrapper for the
  \code{calculateDropout} function.
}
\details{
  Scores dropouts for a dataset.
}

