% Generated by roxygen2 (4.0.0): do not edit by hand
\name{calculateDropout_gui}
\alias{calculateDropout_gui}
\title{Calculate Dropout GUI}
\usage{
calculateDropout_gui(env = parent.frame(), savegui = NULL, debug = FALSE)
}
\arguments{
\item{env}{environment in wich to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}
}
\description{
\code{calculateDropout_gui} is a GUI wrapper for the \code{calculateDropout}
 function.
}
\details{
Scores dropouts for a dataset.
}

