\name{calculateH_gui}
\alias{calculateH_gui}
\title{Calculate H GUI}
\usage{
calculateH_gui(env = parent.frame(), debug = FALSE)
}
\arguments{
  \item{env}{environment in wich to search for data frames
  and save result.}

  \item{debug}{logical indicating printing debug
  information.}
}
\value{
data.frame in slim format.
}
\description{
\code{calculateH_gui} is a GUI wrapper for the
\code{calculateH} function.
}
\details{
Simplifies the use of the \code{calculateH} function by
providing a graphical user interface to it.
}
\references{
Torben Tvedebrink, Poul Svante Eriksen, Helle Smidt
Mogensen, Niels Morling, Evaluating the weight of evidence
by using quantitative short tandem repeat data in DNA
mixtures Journal of the Royal Statistical Society: Series C
(Applied Statistics), Volume 59, Issue 5, 2010, Pages
855-874, 10.1111/j.1467-9876.2010.00722.x.
\url{http://dx.doi.org/10.1111/j.1467-9876.2010.00722.x}
}

