% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateHeterozygous_gui.r
\name{calculateHeterozygous_gui}
\alias{calculateHeterozygous_gui}
\title{Calculate Heterozygous Loci}
\usage{
calculateHeterozygous_gui(env = parent.frame(), debug = FALSE,
  parent = NULL)
}
\arguments{
\item{env}{environment in wich to search for data frames and save result.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI wrapper for the \code{link{calculateHeterozygous}} function.
}
\details{
Simplifies the use of the \code{\link{calculateHeterozygous}} function by
providing a graphical user interface to it.
}
\seealso{
\code{\link{calculateHeterozygous}}
}

