% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calculateHeterozygous_gui}
\alias{calculateHeterozygous_gui}
\title{Calculate Heterozygous}
\usage{
calculateHeterozygous_gui(env = parent.frame(), debug = FALSE,
  parent = NULL)
}
\arguments{
\item{env}{environment in wich to search for data frames and save result.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
\code{calculateHeterozygous_gui} is a GUI wrapper for the
\code{link{calculateHeterozygous}} function.
}
\details{
Simplifies the use of the \code{\link{calculateHeterozygous}} function by
providing a graphical user interface to it.
}
\seealso{
\code{\link{calculateHeterozygous}}
}

