% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateOL_gui.r
\name{calculateOL_gui}
\alias{calculateOL_gui}
\title{Analyze Off-ladder Alleles}
\usage{
calculateOL_gui(
  env = parent.frame(),
  savegui = NULL,
  debug = TRUE,
  parent = NULL
)
}
\arguments{
\item{env}{environment in which to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI wrapper for the \code{\link{calculateOL}} function.
}
\details{
By analysis of the allelic ladder the risk for getting off-ladder
(OL) alleles are calculated. The frequencies from a provided population
database is used to calculate the risk per marker and in total for the given
kit(s). Virtual alleles can be excluded from the calculation.
Small frequencies can be limited to the estimate 5/2N.
}
\seealso{
\code{\link{calculateOL}}
}
