% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePeaks_gui.r
\name{calculatePeaks_gui}
\alias{calculatePeaks_gui}
\title{Calculate Peaks}
\usage{
calculatePeaks_gui(env = parent.frame(), savegui = NULL, debug = FALSE,
  parent = NULL)
}
\arguments{
\item{env}{environment in which to search for data frames.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI wrapper for the \code{\link{calculatePeaks}} function.
}
\details{
Counts the number of peaks in samples and markers with option to
 discard off-ladder peaks and to label groups according to maximum number
 of peaks.
}
\seealso{
\code{\link{calculatePeaks}}
}
