% Generated by roxygen2 (4.0.0): do not edit by hand
\name{calculateResultType_gui}
\alias{calculateResultType_gui}
\title{Calculate Result Type GUI}
\usage{
calculateResultType_gui(env = parent.frame(), savegui = NULL,
  debug = FALSE)
}
\arguments{
\item{env}{environment in wich to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}
}
\description{
\code{calculateResultType_gui} is a GUI wrapper for the \code{calculateResultType}
 function.
}
\details{
Simplifies the use of \code{calculateResultType} by providing a
graphical user interface.
}

