% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateResultType_gui.r
\name{calculateResultType_gui}
\alias{calculateResultType_gui}
\title{Calculate Result Type}
\usage{
calculateResultType_gui(
  env = parent.frame(),
  savegui = NULL,
  debug = FALSE,
  parent = NULL
)
}
\arguments{
\item{env}{environment in which to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI wrapper for the \code{\link{calculateResultType}} function.
}
\details{
Simplifies the use of \code{\link{calculateResultType}} by providing a
graphical user interface.
}
\seealso{
\code{\link{calculateResultType}}
}
