% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSubset.r
\name{checkSubset}
\alias{checkSubset}
\title{Check Subset}
\usage{
checkSubset(data, ref, console = TRUE, ignore.case = TRUE, word = FALSE,
  exact = FALSE, debug = FALSE)
}
\arguments{
\item{data}{a data frame in GeneMapper format containing column 'Sample.Name'.}

\item{ref}{a data frame in GeneMapper format containing column 'Sample.Name', 
OR an atomic vector e.g. a single sample name string.}

\item{console}{logical, if TRUE result is printed to R console,
if FALSE a string is returned.}

\item{ignore.case}{logical, if TRUE case insensitive matching is used.}

\item{word}{logical, if TRUE only word matching (regex).}

\item{exact}{logical, if TRUE only exact match.}

\item{debug}{logical indicating printing debug information.}
}
\description{
Check the result of subsetting
}
\details{
Check if ref and sample names are unique for subsetting.
Prints the result to the R-prompt.
}
\seealso{
\code{\link{grep}}
}
