% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compact.r
\name{compact}
\alias{compact}
\title{Compact Profile}
\usage{
compact(data, per.sample = TRUE, col = "Height", debug = FALSE)
}
\arguments{
\item{data}{data frame with at least 'Sample.Name', 'Marker', 'Allele', and 'Height' columns.}

\item{per.sample}{logical TRUE compact per sample, FALSE for entire dataset (in which case
'Sample.Name' is set to 'Profile').}

\item{col}{character string to specify which column to compact.}

\item{debug}{logical for printing debug information.}
}
\value{
data.frame with columns 'Sample.Name', 'Marker', 'Allele', and 'Height'.
}
\description{
Adds Identical Peaks.
}
\details{
Add peak heights of identical alleles together.
}
\keyword{internal}

