% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cropData_gui.r
\name{cropData_gui}
\alias{cropData_gui}
\title{Crop Or Replace}
\usage{
cropData_gui(env = parent.frame(), savegui = NULL, debug = FALSE,
  parent = NULL)
}
\arguments{
\item{env}{environment in wich to search for data frames.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI simplifying cropping and replacing values in data frames.
}
\details{
Select a data frame from the dropdown and a target column.
Select to discard or replace values and additional options.
Click button to 'Apply' changes.
Multiple actions can be performed on one dataset before saving as
a new dataframe.
NB! Check that datatype is correct before click apply to avoid strange behaviour.
}
\seealso{
\code{\link{trim_gui}}, \code{\link{editData_gui}}, \code{\link{combine_gui}}
}

