% Generated by roxygen2 (4.0.0): do not edit by hand
\name{editData_gui}
\alias{editData_gui}
\title{Edit or view data frames GUI}
\usage{
editData_gui(env = parent.frame(), data = NULL, name = NULL,
  edit = TRUE, debug = FALSE)
}
\arguments{
\item{env}{environment in wich to search for data frames.}

\item{data}{data.frame for instant viewing.}

\item{name}{character string with the name of the provided dataset.}

\item{edit}{logical TRUE for enable edit .}

\item{debug}{logical indicating printing debug information.}
}
\description{
\code{editData_gui} is a GUI simplifying editing and viewing of data frames.
}
\details{
Select a data frame from the dropdown and view/edit. Optionally
save as a new dataframe.
}

