% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editData_gui.r
\name{editData_gui}
\alias{editData_gui}
\title{Edit or View Data Frames}
\usage{
editData_gui(env = parent.frame(), savegui = NULL, data = NULL,
  name = NULL, edit = TRUE, debug = FALSE, parent = NULL)
}
\arguments{
\item{env}{environment in wich to search for data frames.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{data}{data.frame for instant viewing.}

\item{name}{character string with the name of the provided dataset.}

\item{edit}{logical TRUE for enable edit .}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI to edit and view data frames.
}
\details{
Select a data frame from the dropdown and view/edit. Optionally
save as a new dataframe.
}
\seealso{
\code{\link{trim_gui}}, \code{\link{cropData_gui}}, \code{\link{combine_gui}}
}
