% Generated by roxygen2 (4.0.2): do not edit by hand
\name{export_gui}
\alias{export_gui}
\title{Export objects}
\usage{
export_gui(env = parent.frame(), savegui = NULL, debug = FALSE,
  parent = NULL)
}
\arguments{
\item{env}{environment where the objects exist.
Default is the current environment.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
\code{export_gui} is a GUI wrapper for the \code{\link{export}} function.
}
\details{
Simplifies the use of the \code{\link{export}} function by providing a graphical
user interface to it.
}
\seealso{
\code{\link{export}}
}

