% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterProfile_gui.r
\name{filterProfile_gui}
\alias{filterProfile_gui}
\title{Filter Profile}
\usage{
filterProfile_gui(env = parent.frame(), savegui = NULL,
  debug = FALSE, parent = NULL)
}
\arguments{
\item{env}{environment in which to search for data frames.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI wrapper for the \code{\link{filterProfile}} function.
}
\details{
Simplifies the use of the \code{\link{filterProfile}} function
by providing a graphical user interface to it.
All data not matching/matching the reference will be discarded.
Useful for filtering stutters and artifacts from raw typing data or
to identify drop-ins.
}
\seealso{
\code{\link{filterProfile}}, \code{\link{checkSubset}}
}
