% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getDb}
\alias{getDb}
\title{Get allele frequency database}
\usage{
getDb(dbNameOrIndex = NULL, debug = FALSE)
}
\arguments{
\item{dbNameOrIndex}{string or integer specifying the database.
If NULL a vector of available databases is returned.}

\item{debug}{logical indicating printing debug information.}
}
\value{
data.frame with allele frequency database information.
If no matching database or database index is found NA is returned.
If the database was not found NULL is returned.
}
\description{
\code{getDb} gives access to allele frequency databases.
}
\details{
The function provides access to allele frequency databases stored in
the file database.txt in the package directory.
It returns the specified allele frequency database.
}
\examples{
# Show available allele frequency databases.
getDb()
}
\keyword{internal}

