% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStrings.r
\name{getStrings}
\alias{getStrings}
\title{Get Language Strings.}
\usage{
getStrings(language = NA, gui = NA, key = NA)
}
\arguments{
\item{language}{character name of the language.}

\item{gui}{character the function name for the gui to 'translate'.}

\item{key}{character the key to 'translate'. Only used in combination with 'gui'.}
}
\value{
character the retrieved value or NA if not found.
}
\description{
Accepts a language code and gui. Returns the corresponding language strings.
}
\details{
Accepts a language code and gui. Returns the corresponding language strings
for the specified gui function from a text file named as the language code.
}
