% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsave_gui.r
\name{ggsave_gui}
\alias{ggsave_gui}
\title{Save Image}
\usage{
ggsave_gui(ggplot = NULL, name = "", env = parent.frame(),
  savegui = NULL, debug = FALSE, parent = NULL)
}
\arguments{
\item{ggplot}{plot object.}

\item{name}{optional string providing a file name.}

\item{env}{environment where the objects exist.
Default is the current environment.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{object specifying the parent widget to center the message box,
and to get focus when finished.}
}
\value{
TRUE
}
\description{
A simple GUI wrapper for \code{\link{ggsave}}.
}
\details{
Simple GUI wrapper for ggsave.
NB! Uses a workaround bypassing the class check for saving 'complex plots':
Step 1 is performed in the \code{strvalidator} plot functions.
\url{http://stackoverflow.com/a/20433318/2173340}  
Step 2 is performed in this function \code{ggsave_gui}.
\url{http://stackoverflow.com/a/18407452/2173340}
}
\seealso{
\code{\link{ggsave}}
}

