% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_gui.r
\name{import_gui}
\alias{import_gui}
\title{Import Data}
\usage{
import_gui(env = parent.frame(), savegui = NULL, debug = FALSE,
  parent = NULL)
}
\arguments{
\item{env}{environment into which the object will be saved.
Default is the current environment.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI wrapper for the \code{\link{import}} function.
}
\details{
Simplifies the use of the \code{\link{import}} function by providing a graphical
user interface to it.
}
\seealso{
\code{\link{import}}
}
