\name{import_gui}
\alias{import_gui}
\title{Import from GMIDX}
\usage{
  import_gui(env = parent.frame())
}
\arguments{
  \item{env}{environment into which the object will be
  saved. Default is the current environment.}
}
\description{
  \code{import_gui} is a GUI wrapper for the \code{import}
  function.
}
\details{
  Simplifies the use of the \code{import} function by
  providing a graphical user interface to it.
}

