% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plotBalance_gui}
\alias{plotBalance_gui}
\title{Plot Balance GUI}
\usage{
plotBalance_gui(env = parent.frame(), savegui = NULL, debug = FALSE)
}
\arguments{
\item{env}{environment in wich to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}
}
\description{
\code{plotBalance_gui} is a GUI simplifying the creation of plots from balance data.
}
\details{
Plot balance data.
}

