% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotBalance_gui.r
\name{plotBalance_gui}
\alias{plotBalance_gui}
\title{Plot Balance}
\usage{
plotBalance_gui(env = parent.frame(), savegui = NULL, debug = FALSE,
  parent = NULL)
}
\arguments{
\item{env}{environment in wich to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI simplifying the creation of plots from balance data.
}
\details{
Select a dataset to plot and the typing kit used (if not autodetected).
Plot heterozygote peak balance versus the average locus peak height,
the average profile peak height 'H', or by the difference in repeat units
(delta). Plot inter-locus balance versus the average locus peak height, or
the average profile peak height 'H'. Automatic plot titles can be replaced by
custom titles. Sex markers can be excluded. It is possible to plot
logarithmic ratios. A name for the result is automatiaclly suggested.
The resulting plot can be saved as either a plot object or as an image.
}
\seealso{
\url{http://docs.ggplot2.org/current/} for details on plot settings.
}

