% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCapillary_gui.r
\name{plotCapillary_gui}
\alias{plotCapillary_gui}
\title{Plot Capillary Balance}
\usage{
plotCapillary_gui(
  env = parent.frame(),
  savegui = NULL,
  debug = FALSE,
  parent = NULL
)
}
\arguments{
\item{env}{environment in which to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI simplifying the creation of plots from capillary balance data.
}
\details{
Select a dataset to plot from the drop-down menu.
Plot capillary balance as a dotplot, boxplot or as a distribution.
Automatic plot titles can be replaced by custom titles.
A name for the result is automatically suggested.
The resulting plot can be saved as either a plot object or as an image.
}
\seealso{
\url{https://ggplot2.tidyverse.org/} for details on plot settings.
}
