% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistribution_gui.r
\name{plotDistribution_gui}
\alias{plotDistribution_gui}
\title{Plot Distribution}
\usage{
plotDistribution_gui(
  env = parent.frame(),
  savegui = NULL,
  debug = FALSE,
  parent = NULL
)
}
\arguments{
\item{env}{environment in which to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI simplifying the creation of distribution plots.
}
\details{
Plot the distribution of data as cumulative distribution function,
probability density function, or count. First select a dataset, then select
a group (in column 'Group' if any), finally select a column to plot the distribution of.
It is possible to overlay a boxplot and to plot logarithms.
Various smoothing kernels and bandwidths can be specified.
The bandwidth or the number of bins can be specified for the histogram.
Automatic plot titles can be replaced by custom titles.
A name for the result is automatically suggested.
The resulting plot can be saved as either a plot object or as an image.
}
\seealso{
\code{\link{log}}, \code{\link{geom_density}}
}
