% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotDistribution_gui}
\alias{plotDistribution_gui}
\title{Plot Distribution}
\usage{
plotDistribution_gui(env = parent.frame(), savegui = NULL, debug = FALSE,
  parent = NULL)
}
\arguments{
\item{env}{environment in wich to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
\code{plotDistribution_gui} is a GUI simplifying the creation of distribution plots.
}
\details{
Plot the distribution of data as cumulative distribution function or
probability density function. First select a dataset, then select a group
(if any), finally select a column to plot the distribution of.
It is possible to overlay a boxplot.
Various smoothing kernels and bandwidths can be specified.
More info on kernels and bandwidth: \url{http://www.inside-r.org/r-doc/stats/density}
Automatic plot titles can be replaced by custom titles.
A name for the result is automatiaclly suggested.
The resulting plot can be saved as either a plot object or as an image.
}

