\name{plotDropout_gui}
\alias{plotDropout_gui}
\title{Plot Dropout GUI}
\usage{
  plotDropout_gui(env = parent.frame(), debug = FALSE)
}
\arguments{
  \item{env}{environment in wich to search for data frames
  and save result.}

  \item{debug}{logical indicating printing debug
  information.}
}
\description{
  \code{plotDropout_gui} is a GUI simplifying the creation
  of plots from dropout data.
}
\details{
  Plot dropout data.
}

