\name{plotDropout_gui}
\alias{plotDropout_gui}
\title{Plot Dropout GUI}
\usage{
plotDropout_gui(env = parent.frame(), savegui = NULL, debug = FALSE)
}
\arguments{
  \item{env}{environment in wich to search for data frames
  and save result.}

  \item{savegui}{logical indicating if GUI settings should
  be saved in the environment.}

  \item{debug}{logical indicating printing debug
  information.}
}
\description{
\code{plotDropout_gui} is a GUI simplifying the creation of
plots from dropout data.
}
\details{
Plot dropout data.
}

