% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotPeaks_gui}
\alias{plotPeaks_gui}
\title{Plot Peaks}
\usage{
plotPeaks_gui(env = parent.frame(), savegui = NULL, debug = FALSE,
  parent = NULL)
}
\arguments{
\item{env}{environment in wich to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
\code{plotPeaks_gui} is a GUI simplifying the creation of plots from
result type data.
}
\details{
Plot result type data. It is possible to customise titles and font
size. Data can be plotted as as frequency or proportion. The values can be
printed on the plot with custom number of decimals. There are several
colour palettes to chose from.
A name for the result is automatiaclly suggested.
The resulting plot can be saved as either a plot object or as an image.
}

