% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPeaks_gui.r
\name{plotPeaks_gui}
\alias{plotPeaks_gui}
\title{Plot Peaks}
\usage{
plotPeaks_gui(
  env = parent.frame(),
  savegui = NULL,
  debug = FALSE,
  parent = NULL
)
}
\arguments{
\item{env}{environment in which to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI simplifying the creation of plots from result type data.
}
\details{
Plot result type data. It is possible to customize titles and font
size. Data can be plotted as as frequency or proportion. The values can be
printed on the plot with custom number of decimals. There are several
color palettes to chose from.
A name for the result is automatically suggested.
The resulting plot can be saved as either a plot object or as an image.
}
