\name{plotPrecision_gui}
\alias{plotPrecision_gui}
\title{Plot Precision GUI}
\usage{
plotPrecision_gui(env = parent.frame(), savegui = NULL, debug = FALSE)
}
\arguments{
  \item{env}{environment in wich to search for data
  frames.}

  \item{savegui}{logical indicating if GUI settings should
  be saved in the environment.}

  \item{debug}{logical indicating printing debug
  information.}
}
\description{
\code{plotPrecision_gui} is a GUI simplifying the creation
of plots from precision data.
}
\details{
Plot precision data.
}

