% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPullup_gui.r
\name{plotPullup_gui}
\alias{plotPullup_gui}
\title{Plot Pull-up}
\usage{
plotPullup_gui(env = parent.frame(), savegui = NULL, debug = FALSE,
  parent = NULL)
}
\arguments{
\item{env}{environment in wich to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI simplifying the creation of plots from pull-up data.
}
\details{
Select a dataset to plot and the typing kit used (if not autodetected).
Plot pull-up peak ratio versus the peak height of the known allele
Automatic plot titles can be replaced by custom titles.
Sex markers can be excluded. A name for the result is automatiaclly suggested.
The resulting plot can be saved as either a plot object or as an image.
}
\seealso{
\url{http://docs.ggplot2.org/current/} for details on plot settings.
}
