\name{plotResultType_gui}
\alias{plotResultType_gui}
\title{Plot Result Type GUI}
\usage{
plotResultType_gui(env = parent.frame(), savegui = NULL, debug = FALSE)
}
\arguments{
  \item{env}{environment in wich to search for data frames
  and save result.}

  \item{savegui}{logical indicating if GUI settings should
  be saved in the environment.}

  \item{debug}{logical indicating printing debug
  information.}
}
\description{
\code{plotResultType_gui} is a GUI simplifying the creation
of plots from result type data.
}
\details{
Plot result type data.
}

