\name{plotStutter_gui}
\alias{plotStutter_gui}
\title{Plot Stutter GUI}
\usage{
  plotStutter_gui(env = parent.frame(), debug = FALSE)
}
\arguments{
  \item{env}{environment in wich to search for data
  frames.}

  \item{debug}{logical indicating printing debug
  information.}
}
\description{
  \code{plotStutter_gui} is a GUI simplifying the creation
  of plots from stutter data.
}
\details{
  Plot stutter data by parent allele or by peak height.
}

