% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotStutter_gui.r
\name{plotStutter_gui}
\alias{plotStutter_gui}
\title{Plot Stutter}
\usage{
plotStutter_gui(env = parent.frame(), savegui = NULL, debug = FALSE,
  parent = NULL)
}
\arguments{
\item{env}{environment in wich to search for data frames.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI simplifying the creation of plots from stutter data.
}
\details{
Select data to plot in the drop-down menu. Check that the correct
kit has been detected. Plot stutter data by parent allele or by peak height.
Automatic plot titles can be replaced by custom titles.
A name for the result is automatiaclly suggested.
The resulting plot can be saved as either a plot object or as an image.
}
\seealso{
\url{http://docs.ggplot2.org/current/} for details on plot settings.
}

