% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPanelsFile.r
\name{readPanelsFile}
\alias{readPanelsFile}
\title{Read Panels File}
\usage{
readPanelsFile(panel.files, debug = FALSE)
}
\arguments{
\item{panel.files}{string, complete path to Panels file.}

\item{debug}{logical indicating printing debug information.}
}
\value{
data.frame containing the columns 'Panel', 'Marker', 'Color',
'Marker.Min', 'Marker.Max', 'Repeat'.
}
\description{
Reads GeneMapper Panels files.
}
\details{
Reads useful information from Panels files and save it as a data.frame.
}
\keyword{internal}
