\docType{data}
\name{ref11}
\alias{ref11}
\title{ESX17 Positive Control Profile}
\format{A data frame with 17 rows and 4 variables}
\description{
  A dataset in 'GeneMaper' format containing the DNA
  profile of the ESX17 positive control sample with
  homozygotes as two entries.
}
\keyword{datasets}

