% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strvalidator-package.r
\docType{data}
\name{ref3}
\alias{ref3}
\title{ESX17 example data for dropout analysis.}
\format{ASCII text file}
\description{
Reference profiles for source samples.
Text file in GeneMapper format.
}
\keyword{datasets}
