\docType{data}
\name{ref4}
\alias{ref4}
\title{ESX17 example data for dropout analysis.}
\format{A data frame with 98 rows and 3 variables}
\usage{
data(ref4)
}
\description{
A slimmed dataset containing reference profiles for source
samples in set4. Reference 'A2' has douoble entries for
homozygotes. Reference 'F2' has single entries for
homozygotes. Reference 'bc' has douoble entries for
homozygotes, and lower case sample name.
}
\keyword{datasets}

