\docType{data}
\name{set1}
\alias{set1}
\title{Typing data in 'GeneMapper' format}
\format{A data frame with 170 rows and 13 variables}
\usage{
data(set1)
}
\description{
A dataset containing ESX17 genotyping result for 8
replicates of the positive control sample, a negative
control and ladder.
}
\keyword{datasets}

