% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strvalidator-package.r
\docType{data}
\name{set2}
\alias{set2}
\title{SGMPlus example data}
\format{A data frame with 32 rows and 5 variables}
\usage{
data(set2)
}
\description{
A slimmed dataset containing SGM Plus genotyping result for 2 replicates
of 'sampleA'.
}
\keyword{datasets}

