% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strvalidator-package.r
\docType{data}
\name{set3}
\alias{set3}
\title{ESX17 example data for dropout analysis.}
\format{
ASCII text file
}
\description{
Data from dilution experiment for dropout analysis.
Text file with exported GeneMapper genotypes table.
}
\keyword{datasets}
