% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strvalidator-package.r
\docType{data}
\name{set5}
\alias{set5}
\title{ESX17 example data for mixture analysis.}
\format{A data frame with 1663 rows and 7 variables}
\usage{
data(set5)
}
\description{
A slimmed dataset containing data from
mixture experiment for Mx analysis.
}
\keyword{datasets}

