% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim.r
\name{slim}
\alias{slim}
\title{Slim Data Frames}
\usage{
slim(data, fix = NULL, stack = NULL, keep.na = TRUE, debug = FALSE)
}
\arguments{
\item{data}{data.frame.}

\item{fix}{vector of strings with colum names to keep fixed.}

\item{stack}{vector of strings with colum names to slim.}

\item{keep.na}{logical, keep a row even if no data.}

\item{debug}{logical indicating printing debug information.}
}
\value{
data.frame
}
\description{
Slim data frames with repeated columns.
}
\details{
Stack repeated columns into single columns.
For example, the following data frame:
 Sample.Name|Marker|Allele.1|Allele.2|Size.1|Size.2|Data.Point..
using this command:
 slim(data, fix=c("Sample.Name","Marker"), stack=c("Allele","Size"))
would result in this data frame (NB! 'Data.Point' is dropped):
 Sample.Name|Marker|Allele|Size
}
