\docType{package}
\name{strValidator-package}
\alias{strValidator}
\alias{strValidator-package}
\title{Simplifies internal validation of forensic STR typing kits}
\description{
Simplifies internal validation of forensic STR typing kits
}
\author{
Oskar Hansson \email{oskar.hansson@fhi.no}
}
\keyword{package}

