\docType{package}
\name{strValidator-package}
\alias{strValidator}
\alias{strValidator-package}
\title{Simplifies internal validation of forensic STR typing kits}
\description{
  Internal validation of forensic STR kits made easy with
  strValidator.
}
\details{
  strValidator provide tools that make it easier to perform
  an internal validation of a forensic short tandem repeat
  (STR) kit for human identification. The tools are
  developed to provide all necessary data to conform with
  the guidelines in REF... More information about each
  function can be found in its help documentation.

  \tabular{ll}{ Package: \tab strValidator\cr Type: \tab
  Package\cr Version: \tab 0.1\cr Date: \tab 2013-02-06\cr
  License: \tab GPL (>= 2)\cr LazyLoad: \tab yes\cr }
}
\section{Warning}{
  This package is experimental and has not been thoroughly
  validated.
}
\author{
  Oskar Hansson \email{oskar.hansson@fhi.no}
}
\keyword{package}

