% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strvalidator-package.r
\docType{package}
\name{strValidator-package}
\alias{strValidator}
\alias{strValidator-package}
\title{Simplifies internal validation of forensic STR typing kits}
\author{
Oskar Hansson \email{oskar.hansson@fhi.no}
}
\keyword{package}

