% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strvalidator.r
\name{strvalidator}
\alias{strvalidator}
\title{Graphical User Interface For The STR-validator Package}
\usage{
strvalidator(debug = FALSE)
}
\arguments{
\item{debug}{logical indicating printing debug information.}
}
\value{
TRUE
}
\description{
Validation and process control toolbox for forensic genetic laboratories.
}
\details{
STR-validator (pronounced starvalidator) is a package developed
for validation and process control of both methods and instruments in a
forensic genetic laboratory setting.
This graphical user interface make it very easy to analyse validation
data in accordance with ENFSI and SWGDAM guidelines.
The code has been extensively tested in order to assure correct results.\cr\cr

Open the help page and click \code{Index} at the bottom of the page
to see a list of all functions.\cr\cr

Created by:\cr
Oskar Hansson, Department of Forensic Biology (NIPH, Norway)\cr\cr

General user information and tutorials:\cr
\url{https://sites.google.com/site/forensicapps/strvalidator}\cr\cr

Facebook user community:\cr
\url{https://www.facebook.com/pages/STR-validator/240891279451450?ref=tn_tnmn}\cr\cr
\url{https://www.facebook.com/groups/strvalidator/}\cr\cr

Please report bugs to:\cr
\url{https://github.com/OskarHansson/strvalidator/issues}\cr\cr

The source code and collaborative community is hosted at GitHub:\cr
\url{https://github.com/OskarHansson/strvalidator}\cr\cr
}
\examples{
# To start the graphical user interface.
\dontrun{
strvalidator()
}
}
\references{
Recommended Minimum Criteria for the Validation of Various Aspects of the DNA Profiling Process
(access have been unstable lately but it is usually found at http://www.enfsi.eu/sites/default/files/documents/minimum_validation_guidelines_in_dna_profiling_-_v2010_0.pdf)
Validation Guidelines for Forensic DNA Analysis Methods (2012)
\url{http://swgdam.org/SWGDAM_Validation_Guidelines_APPROVED_Dec_2012.pdf}
}

