% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strvalidator.r
\name{strvalidator}
\alias{strvalidator}
\title{Graphical User Interface For The STR-validator Package}
\usage{
strvalidator(debug = FALSE)
}
\arguments{
\item{debug}{logical indicating printing debug information.}
}
\value{
TRUE
}
\description{
GUI simplifying the use of the strvalidator package.
}
\details{
The graphical user interface give easy access to all graphical
versions of the functions available in the strvalidator package. It connects
functions 'under the hood' to allow a degree of automation not available
using the command based functions. In addition it provides a project based
workflow.\cr\cr
Click \code{Index} at the bottom of the help page to see a complete list
of functions.
}
\examples{
# To start the graphical user interface.
\dontrun{
strvalidator()
}

}
