% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableBalance_gui.r
\name{tableBalance_gui}
\alias{tableBalance_gui}
\title{Table Balance}
\usage{
tableBalance_gui(
  env = parent.frame(),
  savegui = NULL,
  debug = FALSE,
  parent = NULL
)
}
\arguments{
\item{env}{environment in which to search for data frames.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
GUI wrapper for the \code{\link{tableBalance}} function.
}
\details{
Simplifies the use of the \code{\link{tableBalance}} function by providing a graphical
user interface to it.
}
\seealso{
\code{\link{tableBalance}}
}
