% Generated by roxygen2 (4.0.0): do not edit by hand
\name{tableCapillary_gui}
\alias{tableCapillary_gui}
\title{Guess profile}
\usage{
tableCapillary_gui(env = parent.frame(), savegui = NULL, debug = FALSE)
}
\arguments{
\item{env}{environment in wich to search for data frames.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}
}
\description{
\code{tableCapillary_gui} is a GUI wrapper for the \code{tableCapillary} function.
}
\details{
Simplifies the use of the \code{tableCapillary} function by providing a graphical
user interface to it.
}

