% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tablePrecision.r
\name{tablePrecision}
\alias{tablePrecision}
\title{Calculate Precision}
\usage{
tablePrecision(data, key = c("Marker", "Allele"), target = c("Size"),
  debug = FALSE)
}
\arguments{
\item{data}{Data frame containing at least columns defined in 'key' and 'target'.}

\item{key}{vector containing column names to create keys from.}

\item{target}{vector containing column <base> names to calculate precision for.}

\item{debug}{logical indicating printing debug information.}
}
\value{
data.frame with results.
}
\description{
Summarize precision analysis result in table format.
}
\details{
Calculates summary statistics for 'target' columns for each unique
'key' combination. For example the precision of determined size for alleles
in multiple allelic ladders.
Requires a 'slimmed' and 'filtered' data frame.
For more details see \code{min}, \code{max}, \code{mean}, \code{sd}, \code{quantile}.
}

