% Generated by roxygen2 (4.0.0): do not edit by hand
\name{tableStutter_gui}
\alias{tableStutter_gui}
\title{Table stutter}
\usage{
tableStutter_gui(env = parent.frame(), savegui = NULL, debug = FALSE)
}
\arguments{
\item{env}{environment in wich to search for data frames.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}
}
\description{
\code{tableStutter_gui} is a GUI wrapper for the \code{tableStutter} function.
}
\details{
Simplifies the use of the \code{tableStutter} function by providing a graphical
user interface to it.
}

