% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tableStutter_gui}
\alias{tableStutter_gui}
\title{Table Stutter}
\usage{
tableStutter_gui(env = parent.frame(), savegui = NULL, debug = FALSE,
  parent = NULL)
}
\arguments{
\item{env}{environment in wich to search for data frames.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
\code{tableStutter_gui} is a GUI wrapper for the \code{\link{tableStutter}} function.
}
\details{
Simplifies the use of the \code{\link{tableStutter}} function by providing a graphical
user interface to it.
}
\seealso{
\code{\link{tableStutter}}
}

