% Generated by roxygen2 (4.0.0): do not edit by hand
\name{trim_gui}
\alias{trim_gui}
\title{Trim data GUI}
\usage{
trim_gui(env = parent.frame(), savegui = NULL, debug = FALSE)
}
\arguments{
\item{env}{environment in wich to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}
}
\value{
data.frame with extracted result.
}
\description{
\code{trim_gui} is a GUI wrapper for the \code{trim} function.
}
\details{
Simplifies the use of the \code{trim} function by providing a graphical
user interface to it.
}

