% Generated by roxygen2 (4.0.1): do not edit by hand
\name{trim_gui}
\alias{trim_gui}
\title{Trim data}
\usage{
trim_gui(env = parent.frame(), savegui = NULL, debug = FALSE,
  parent = NULL)
}
\arguments{
\item{env}{environment in wich to search for data frames and save result.}

\item{savegui}{logical indicating if GUI settings should be saved in the environment.}

\item{debug}{logical indicating printing debug information.}

\item{parent}{widget to get focus when finished.}
}
\value{
TRUE
}
\description{
\code{trim_gui} is a GUI wrapper for the \code{\link{trim}} function.
}
\details{
Simplifies the use of the \code{\link{trim}} function by providing a graphical
user interface to it.
}
\seealso{
\code{\link{trim}}
}

